Give windows a little more time to startup
The default waitFor timeout is 1 second which is not enough on CI
